+++ src/Analyzers.CodeFixes/CSharp/CodeFixes/EnumMemberDeclarationCodeFixProvider.cs
@@ -192,6 +192,9 @@ public sealed class EnumMemberDeclarationCodeFixProvider : BaseCodeFixProvider
     {
         foreach (EnumMemberDeclarationSyntax enumMember in enumDeclaration.Members)
         {
+            if (CSharpUtility.IsEnumMemberObsolete(enumMember))
+                continue;
+
             IFieldSymbol fieldSymbol = semanticModel.GetDeclaredSymbol(enumMember, cancellationToken);
 
             if (!SymbolEqualityComparer.Default.Equals(fieldSymbolInfo.Symbol, fieldSymbol))