+++ src/Tests/Analyzers.Tests/RCS1234DuplicateEnumValueTests.cs
@@ -115,6 +115,34 @@ enum E
     }
 
     [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.DuplicateEnumValue)]
+    public async Task Test6()
+    {
+        await VerifyDiagnosticAndFixAsync(@"
+using System;
+
+enum E
+{
+    [Obsolete]
+    A = 1,
+
+    B = 1,
+    C = [|1|]
+}
+", @"
+using System;
+
+enum E
+{
+    [Obsolete]
+    A = 1,
+
+    B = 1,
+    C = B
+}
+");
+    }
+
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.DuplicateEnumValue)]
     public async Task TestNoDiagnostic()
     {
         await VerifyNoDiagnosticAsync(@"
@@ -123,5 +151,21 @@ enum E
     A,
     B = A
 }");
+    }
+
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.DuplicateEnumValue)]
+    public async Task TestNoDiagnostic2()
+    {
+        await VerifyNoDiagnosticAsync(@"
+using System;
+
+enum E
+{
+    [Obsolete]
+    A = 1,
+
+    B = 1,
+}
+");
     }
 }