+++ src/CommandLine/Commands/FixCommand.cs
@@ -45,8 +45,6 @@ internal class FixCommand : MSBuildWorkspaceCommand<FixCommandResult>
 
     public override async Task<FixCommandResult> ExecuteAsync(ProjectOrSolution projectOrSolution, CancellationToken cancellationToken = default)
     {
-        AssemblyResolver.Register();
-
         var codeFixerOptions = new CodeFixerOptions(
             fileSystemFilter: FileSystemFilter,
             severityLevel: SeverityLevel,