+++ ChangeLog.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix [RCS1234](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1234) ([PR](https://github.com/dotnet/roslynator/pull/1233))
 - Fix refactoring [Inline method](https://josefpihrt.github.io/docs/roslynator/refactorings/RR0062) ([PR](https://github.com/dotnet/roslynator/pull/1234))
 
+### Added
+- Tests for [RCS1196](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1196/)
+
 ## [4.6.1] - 2023-10-23
 
 ### Fixed