+++ ChangeLog.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix refactoring [Inline method](https://josefpihrt.github.io/docs/roslynator/refactorings/RR0062) ([PR](https://github.com/dotnet/roslynator/pull/1234))
 
 ### Added
-- Tests for [RCS1196](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1196/)
+- Tests for [RCS1196](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1196/) ([PR](https://github.com/dotnet/roslynator/pull/1235))
 
 ## [4.6.1] - 2023-10-23