+++ ChangeLog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix [RCS1234](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1234) ([PR](https://github.com/dotnet/roslynator/pull/1233))
 - Fix refactoring [Inline method](https://josefpihrt.github.io/docs/roslynator/refactorings/RR0062) ([PR](https://github.com/dotnet/roslynator/pull/1234))
-- [CLI] Remove assembly resolving ([PR](https://github.com/dotnet/roslynator/pull/1237))
+- [CLI] Remove assembly resolving ([PR](https://github.com/dotnet/roslynator/pull/1237)) 
 
 ## [4.6.1] - 2023-10-23